requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
auth0-python==3.16.1
backcall==0.2.0
bios==0.1.2
black==21.12b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==7.1.2
cryptography==36.0.1
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.3
executing==0.8.2
h11==0.12.0
idna==2.10
ipykernel==6.6.1
ipython==8.0.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.0.1
matatika==1.11.0
matatika-iplotter==1.2.1
matplotlib-inline==0.1.3
mdutils==1.3.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
outcome==1.1.0
oyaml==1.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.1
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.14.0
PyYAML==5.3.1
pyzmq==22.2.1
requests==2.24.0
selenium==3.9.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
stack-data==0.1.3
terminado==0.12.1
testpath==0.5.0
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.1
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
wsproto==1.0.0
